pyproject.toml
[tool.poetry]
name = "notifrw"
version = "0.1.0"
description = ""
authors = ["Igor Pavlyukovets <igor.p.i@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
python-telegram-bot = "^21.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
